Moved Packages class definition to seperate file
